% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrivalFixingLoading.R
\name{consequential_mTable}
\alias{consequential_mTable}
\title{consequential_mTable}
\usage{
consequential_mTable(mTable)
}
\arguments{
\item{mTable}{master Table generated by createMT() function}
}
\value{
filtered data.table / tibble
}
\description{
selects columns which are useful for graphing. Also filters out variants which which are inconsequential according to their EnsVar_most_severe_consequence
}
\details{
unfiltered master table goes in, narrowed down table comes out
}
