% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugging_APIcall_functions.R
\name{createMT2}
\alias{createMT2}
\title{createMT2}
\usage{
createMT2(
  fileFolderPath,
  varAnnotations = TRUE,
  population_data = FALSE,
  processData = TRUE
)
}
\arguments{
\item{fileFolderPath}{NA}

\item{varAnnotations}{NA}
}
\value{
NA
}
\description{
debug func
}
\details{
debug func
}
\examples{
NA


}
