% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugging_APIcall_functions.R
\name{dbugTransformOPT}
\alias{dbugTransformOPT}
\title{dbugTransformOPT}
\usage{
dbugTransformOPT(dataList, popsData = F)
}
\arguments{
\item{fileFolderPath}{NA}

\item{varAnnotations}{NA}
}
\value{
NA
}
\description{
debug func
}
\details{
debug func
}
\examples{
NA


}
