% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrivalFixingLoading.R
\name{ensListTransform}
\alias{ensListTransform}
\title{ensListTransform}
\usage{
ensListTransform(dataList, popsData = F)
}
\arguments{
\item{dataList}{Data to be transformed. Must be in list format such that a GWAS data table is the first element and the respective data produced from calling Ensembl's REST API Variants endpoint with get_ensVariants() is the second element within the list. Position in the list is critical to successful execution of this function}

\item{popsData}{populations data transformation option. when TRUE function runs assuming variants from Ensembl REST API have been called with populations option activated, resulting output is different due to this extra population data.}
}
\value{
data.frame or list of data.frames
}
\description{
Transforms list-form data which is produced by \code{get_ensVariants()} and \code{createMT(processData = FALSE)} into flat a list of flat tables which can then be used for graphing or viewing data in tabular form.
}
\examples{
NA

}
