% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugging_APIcall_functions.R
\name{get_ensVariants2}
\alias{get_ensVariants2}
\title{get_ensVariants2}
\usage{
get_ensVariants2(rsIDs, population_data = FALSE, processData = TRUE)
}
\arguments{
\item{rsIDs}{vector of rsIDs (variant IDs of the rs00000000 form)}

\item{population_data}{when TRUE, activates the option to grab population data for each variant in the request as well.}
}
\value{
list of data.frame(s) / data.table(s)
}
\description{
debug func
}
\details{
debug func
}
\examples{
variantData <- get_ensVariants(c("rs11137048","rs6866110", "rs62227671", "rs6122625", "rs57504074"), population_data = TRUE)


}
