% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_graphing.R
\name{graph_singlePopAlleleFreq}
\alias{graph_singlePopAlleleFreq}
\title{graph_singlePopAlleleFreq}
\usage{
graph_singlePopAlleleFreq(
  masterTable,
  singlePopTable,
  numVariants = nrow(preparedData),
  graph_style = "point",
  facet_graph = FALSE,
  orderless = FALSE,
  ascending = FALSE,
  show_genes = FALSE,
  pValAsSize = FALSE,
  export = FALSE,
  file_name = "SinglePopFreqGraph",
  graph_scale = 3,
  sqrtYscale = FALSE,
  yUpperLim = NA
)
}
\arguments{
\item{masterTable}{Data in. - Table of SNP data from GWAS catalog and Ensembl-variants endpoint. Output by get_ensVariants(). Contains 40 variables.}

\item{singlePopTable}{Data in - A population allele / SNP frequency table which is produced by singlePopTransform() function. Contains 'VariantID' 'allele_count' 'frequency' 'allele' variables, as well as a population attribute}

\item{numVariants}{specifys the number of variants to be displayed by the graph.}

\item{graph_style}{can be either 'bar' / 'col' or 'point' .. changes the geometry of the graph, the two different styles convey some different information as well. (Try them both!) (bar and col are synonymous and produce the same graph)}

\item{facet_graph}{when TRUE, graph will be faceted according to 'Reported_trait'.}

\item{orderless}{when TRUE, the graph will be unordered and thus bars / points will be randomly distributed instead of sorted}

\item{ascending}{if TRUE the graphs will display in acsending order rather than descending order (which is the default behavior)}

\item{show_genes}{displays gene names of each variants above their respective data point on the graph.}

\item{pValAsSize}{changes point sizes according to p-value of each SNP, ONLY FUNCTIONS for 'point' graph_style}

\item{export}{determines whether or not to export the graph generated}

\item{file_name}{allows the user to specify the name of the file output if 'export' is set to TRUE}

\item{graph_scale}{determines the scale of the graph for exported files, larger values generate higher resolution graphs}

\item{sqrtYscale}{When TRUE, very small frequency values are made visible by elongating the lower range of the y-axis. sqrtYscale option will over-ride the 'yUpperLim' option}

\item{yUpperLim}{sets the y-axis to a custom value, by default graphing function ggplot() automatically determines the y-axis limit; yUpperLim will override sqrtYscale if both are active}
}
\value{
ggplot2 graph object
}
\description{
Graphing function which takes in a master table and a single-population allele frequency table and outputs a customizable graph which is either 'point' or 'bar' geometry / style.
}
\details{
See 'Arguments' for details on graph customization / options. It is strongly suggested to play around with the various optional arguments in order to discover the range of possible outputs when seeking any specific look or style of graph.
}
\examples{
# using package data object 'testMasterList' for this example
   testMasterTable <- testMasterList[[1]]
   singlePopulation <- testMasterList[[3]][[1]]

# graphing with default options, see vignettes at https://github.com/J-T-Nelson/GWASpops.pheno2geno for more details on graphs which this function can produce
   graph_singlePopAlleleFreq(testMasterTable, singlePopulation)

}
