% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_graphing.R
\name{graph_varPopFrequencies}
\alias{graph_varPopFrequencies}
\title{graph_varPopFrequencies}
\usage{
graph_varPopFrequencies(
  popFreqTable,
  numPopulations = nrow(preparedData),
  graph_style = "bar",
  orderless = FALSE,
  ascending = FALSE,
  export = FALSE,
  file_name = "PopFrequencyGraph",
  graph_scale = 3,
  sqrtYscale = FALSE,
  yUpperLim = NA
)
}
\arguments{
\item{numPopulations}{is the number of populations shown on the graph. If it is not adjusted by the user, the graph is unsorted by default, if the user changes the value of numPopulations the data graphed will be sorted descending by default, but can be adjusted to ascending by using the 'ascending' option.}

\item{graph_style}{can be either 'bar' or 'point' .. changes the geometry of the graph, the two different styles convey some different information as well. (Try them both!)}

\item{orderless}{if TRUE, will overide 'ascending' and enforce orderlessness upon the x axis elements}

\item{ascending}{if TRUE the graphs will display in acsending order rather than descending order (which is the default behavior)}

\item{export}{determines whether or not to export the graph generated}

\item{file_name}{allows the user to specify the name of the file output if 'export' is set to TRUE}

\item{graph_scale}{determines the scale of the graph for exported files, larger values generate higher resolution graphs}

\item{sqrtYscale}{When TRUE, very small frequency values are made visible by elongating the lower range of the y-axis. sqrtYscale option will over-ride the 'yUpperLim' option}

\item{yUpperLim}{sets the y-axis to a custom value, by default graphing function ggplot() automatically determines the y-axis limit.}
}
\value{
ggplot2 graph object
}
\description{
graphing function which takes in a population allele frequency table for a single variant (rs00000000), and outputs a customizable graph of either 'bar' or 'point' geometry-styles.
}
\details{
See params for details on graph customization / options. It is strongly suggested to play around with the various optional arguments in order to discover the range of possible outputs when seeking any specific look or style of graph.
popFreqTable's are the data frames which are generated by createMT(), get_ensVariants(). The tables are named after the Variant which they contain data for. They contain 'allele', 'population', 'allele_count' and 'frequency' variables
}
