% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWAS.asso.study.data}
\alias{GWAS.asso.study.data}
\title{GWAS Catalog Associations and Studies Tables Merged}
\format{
\subsection{\code{GWAS.asso.study.data}}{

A data.table with 202 rows and 22 columns, the columns / variables come directly from GWAS catalog information concerning the meaning of each variable can be found at \url{https://www.ebi.ac.uk/gwas/docs/faq}. Additional searching within the GWAS catalog site may be necessary to verify exact meaning and purpose of values within this data set.
}
}
\source{
\url{https://www.ebi.ac.uk/gwas/search?query=air\%20pollution}
}
\usage{
GWAS.asso.study.data
}
\description{
After downloading studies and associations tables from GWAS catalog for several terms which fall under the umbrella 'air pollution' they were merged into a single table.
The data is all generated by researchers performing GWAS studies, and curated by the maintainers of the GWAS catalog.
}
\keyword{datasets}
