% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugging_APIcall_functions.R
\name{multiAPIcall_variants2}
\alias{multiAPIcall_variants2}
\title{multiAPIcall_variants2}
\usage{
multiAPIcall_variants2(rsIDs, popData = FALSE, procsData = TRUE)
}
\arguments{
\item{fileFolderPath}{NA}

\item{varAnnotations}{NA}
}
\value{
NA
}
\description{
debug func
}
\details{
debug func
}
\examples{
NA


}
