% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Debugging_APIcall_functions.R
\name{multiAPIcall_variantsFinal}
\alias{multiAPIcall_variantsFinal}
\title{multiAPIcall_variantsFinal}
\usage{
multiAPIcall_variantsFinal(rsIDs, popData = FALSE)
}
\arguments{
\item{fileFolderPath}{NA}

\item{varAnnotations}{NA}
}
\value{
NA
}
\description{
debug func
}
\details{
debug func
}
\examples{
NA


}
