% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Populations}
\alias{Populations}
\title{Populations of Genomic Data - Ensembl Aggregation}
\format{
\subsection{\code{Populations}}{

A data.table with 73 rows and 4 columns:
\describe{
\item{Population_abbreviation}{The abbreviated terms which describe the project source of data as well as the population sampled in the project}
\item{Sample_Count}{Number of individuals sampled}
\item{Pop_Ancestry}{Ancestry of each population}
\item{PopAnces_Graph_Labels}{Condensed versions of 'Pop_Ancestry' column, created for use as labels when graphing. The shortened names capture the most pertinent details of each ancestry while allowing for cleaner more uniform labeling in graphs.}
}
}
}
\source{
\url{https://useast.ensembl.org/info/genome/variation/species/populations.html}
}
\usage{
Populations
}
\description{
Key population labels which map projects and their specific terms for populations to human readable population ancestries. Sample counts of various populations is also included.
}
\keyword{datasets}
