% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testMasterList}
\alias{testMasterList}
\title{Master List for package examples}
\format{
\subsection{\code{testMasterList}}{

A list with 3 elements. 1. masterTable, 2. PopAlleleFreqData 3. singlePop_alleleFreqDTs
}
}
\source{
\url{https://www.ebi.ac.uk/gwas/search?query=air\%20pollution}, \url{https://useast.ensembl.org/info/genome/variation/species/populations.html}
}
\usage{
testMasterList
}
\description{
full data list which can be used to learn the package.
Also serves as an example to verify the data you get from calling createMT() is coming out in proper form with all desired variables.
}
\details{
This master list uses the same base data as the 'GWAS.asso.study.data' data object which comes with the package as well, but also includes data from Ensembl.
}
\keyword{datasets}
