% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrivalFixingLoading.R
\name{write_csv_listOf_DFs}
\alias{write_csv_listOf_DFs}
\title{write_csv_listOf_DFs}
\usage{
write_csv_listOf_DFs(
  listOfDataFrames,
  destinationFolder = NA,
  pattern = NA,
  replacementChar = NA,
  sortTables = FALSE,
  sortingCriteria = "variableName",
  ascending = FALSE
)
}
\arguments{
\item{destinationFolder}{folder the user wishes to export files to. By default files will export to current working directory}

\item{pattern}{specifies character or characters within data table names to replace by 'replacementChar'}

\item{replacementChar}{replacement character which will be inserted in place of 'patten' within file names}

\item{sortTables}{when TRUE will sort the tables according to the column specified by 'sortingCriteria'}

\item{sortingCriteria}{name of column which will be used to sort the data table by.}

\item{ascending}{will change the default sorting order from descending to ascending}
}
\value{
NA
}
\description{
iterates over a list of data.frames or table like data structures writing each to their own csv files.
}
\details{
Names of files come from the names of the tables within the list input. Destination folder can be selected otherwise files will be exported into current wd.
}
