% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensVariants.R
\name{get_ensVariants}
\alias{get_ensVariants}
\title{get_ensVariants}
\usage{
get_ensVariants(rsIDs, population_data = FALSE)
}
\arguments{
\item{rsIDs}{vector of rsIDs (variant IDs of the rs00000000 form)}

\item{population_data}{when TRUE, activates the option to grab population data for each variant in the request as well.}
}
\value{
list of data.frame(s) / data.table(s)
}
\description{
grabs ensVariants from API and converts to tabular form. Optionally grabs population data as well.
}
\details{
Primary method to grab variant data from Ensembl REST API in the pipeline.

Use this function if you're interested in grabbing data on specific variants from Ensembl's Variants endpoint.

Calls the variants POST endpoint, build into the function are several layers of processing which both ensure arbitrarily large calls are executed successfully, as well as that data returned is in a set of tabular semi-flat formatted objects.

Requesting population data will substantially increase the time this function takes to complete due to the increase in data being transferred.
}
\examples{
variantData <- get_ensVariants(c("rs11137048","rs6866110", "rs62227671", "rs6122625", "rs57504074"), population_data = TRUE)

}
