% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_retrivalFixingLoading.R
\name{write_csv_listOf_DFs}
\alias{write_csv_listOf_DFs}
\title{write_csv_listOf_DFs}
\usage{
write_csv_listOf_DFs(
  listOfDataFrames,
  destinationFolder = NA,
  pattern = NA,
  replacementChar = NA,
  sortTables = FALSE,
  sortingCriteria = "variableName",
  ascending = FALSE
)
}
\arguments{
\item{listOfDataFrames}{Data in. A list of data.frame(s) or similarly shaped objects which can be exported as CSVs.}

\item{destinationFolder}{folder the user wishes to export files to. By default files will export to current working directory}

\item{pattern}{specifies character or characters within data table names to replace by 'replacementChar'}

\item{replacementChar}{replacement character which will be inserted in place of 'pattern' within file names}

\item{sortTables}{when TRUE will sort the tables according to the column specified by 'sortingCriteria'}

\item{sortingCriteria}{name of column which will be used to sort the data table by.}

\item{ascending}{will change the default sorting order from descending to ascending}
}
\description{
iterates over a list of data.frames or table like data structures writing each to their own csv file.
}
\details{
Names of files come from the names of the tables within the list input. Destination folder can be selected otherwise files will be exported into current wd.
}
\examples{
# use internal example masterList for this example
     masterList <- testMasterList

# grabbing just a few tables for examples sake
     tableList <- masterList[[2]][1:5]

# creating a single csv for each of the 5 tables, replacing 'rs' in the name with 'Table_Prefix',
# they're being created in the current working directory since no destination folder is specified.
     write_csv_listOf_DFs(tableList, pattern = 'rs', replacementChar = 'Table_Prefix')


}
